requirements.txt
pandas
numpy
scikit-learn
matplotlib
plotly
seaborn
jupyter
pyyaml
pytest
